refactor: add eslint config and change code when using command npx config --fix
