--- create.js
@@ -1,8 +1,0 @@
-const createElement = (element, innerHTML, ...classes) => {
-    const node = document.createElement(element);
-    node.classList.add(...classes);
-    node.innerHTML = innerHTML;
-    return node;
-};
-
-export default createElement;
