+++ index.html
@@ -10,11 +10,6 @@
   <link rel="stylesheet" href="style.css">
 </head>
 <body>
-
-
-
-
+<script type="module" src="js/action.js"></script>
 </body>
-<script type="module" src="action.js"></script>
-
 </html>
