+++ js/create.js
@@ -1,0 +1,8 @@
+const createElement = (element, innerHTML, ...classes) => {
+  const node = document.createElement(element);
+  node.classList.add(...classes);
+  node.innerHTML = innerHTML;
+  return node;
+};
+
+export default createElement;
